addGroupModal を別コンポーネントとして切り出し #50
