+++ backend/resources/ts/components/Group.tsx
@@ -1,17 +1,12 @@
 import React, { useCallback, useEffect, useState } from 'react'
 import { Link } from 'react-router-dom'
-import Modal from 'react-modal'
 import {
   useAppDispatch,
   useGroupsEntities,
   useGroupsIds,
-  useGroupsPromise,
-  useModalStyle,
   useParamGroupId,
 } from '../app/hooks'
-import { addGroup, fetchGroups, updateGroups } from '../slices/GroupsSlice'
-
-if (process.env.NODE_ENV !== 'test') Modal.setAppElement('#app')
+import AddGroupModal from './AddGroupModal'
 
 const Group = () => {
   const dispatch = useAppDispatch()
@@ -63,60 +58,6 @@ const GroupItem = React.memo(({ id, name }: GroupItemProps) => {
           <span className="ml-2">{name}</span>
         </Link>
       </li>
-    </>
-  )
-})
-
-const AddGroupModal = React.memo(() => {
-  const [isOpen, setIsOpen] = useState(false)
-  const [groupName, setGroupName] = useState('')
-  const dispatch = useAppDispatch()
-  const modalStyle = useModalStyle()
-
-  const addGroupHandler = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault()
-    dispatch(addGroup({ groupName: groupName }))
-    setGroupName('')
-    console.log('aadd')
-
-    closeModal()
-  }
-  const onChangeNameHandler = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setGroupName(e.target.value)
-  }
-
-  const openModal = () => {
-    setIsOpen(true)
-  }
-  const closeModal = () => {
-    setIsOpen(false)
-  }
-
-  return (
-    <>
-      <button className="btn btn-primary" onClick={openModal}>
-        新しいグループを追加
-      </button>
-      <Modal isOpen={isOpen} onRequestClose={closeModal} style={modalStyle}>
-        <h4>追加するグループの名前</h4>
-        <form onSubmit={addGroupHandler} className="form">
-          <input
-            type="text"
-            name="groupName"
-            id="groupName"
-            className="from-controll"
-            value={groupName || ''}
-            onChange={onChangeNameHandler}
-            autoFocus
-          />
-          <button className="btn btn-primary" type="submit">
-            追加
-          </button>
-          <button className="btn btn-light" onClick={closeModal}>
-            キャンセル
-          </button>
-        </form>
-      </Modal>
     </>
   )
 })
